+++ app/page.tsx
@@ -1,203 +1,48 @@
 "use client"
-import { useState, useEffect } from 'react'
+import { useState } from 'react'
+import { useRouter } from 'next/navigation'
 
-interface CostItem {
-  id: string
-  category: string
-  label: string
-  cost: number
-  isEditing?: boolean
-}
+export default function Login() {
+  const [username, setUsername] = useState('')
+  const [password, setPassword] = useState('')
+  const router = useRouter()
 
-interface Empanada {
-  name: string
-  costs: CostItem[]
-  margin: number
-}
-
-const initialCosts = [
-  { id: 'carne', category: 'Relleno', label: 'Carne', cost: 1.95 },
-  { id: 'cebolla', category: 'Relleno', label: 'Cebolla', cost: 0.8 },
-  { id: 'azafran', category: 'Relleno', label: 'Azafrán', cost: 0.025 },
-  { id: 'sal_relleno', category: 'Relleno', label: 'Sal', cost: 0.0048 },
-  { id: 'azucar', category: 'Relleno', label: 'Azúcar', cost: 0.012 },
-  { id: 'aceite_relleno', category: 'Relleno', label: 'Aceite', cost: 0.08 },
-  { id: 'gas_cocina', category: 'Relleno', label: 'Gas cocina', cost: 0.026 },
-
-  { id: 'harina', category: 'Masa', label: 'Harina', cost: 0.112 },
-  { id: 'aceite_masa', category: 'Masa', label: 'Aceite', cost: 0.088 },
-  { id: 'margarina', category: 'Masa', label: 'Margarina vegetal', cost: 0.031 },
-  { id: 'sal_masa', category: 'Masa', label: 'Sal', cost: 0.0008 },
-  { id: 'levadura', category: 'Masa', label: 'Levadura', cost: 0.0088 },
-  { id: 'electricidad_amasadora', category: 'Masa', label: 'Electricidad amasadora', cost: 0.0358 },
-
-  { id: 'bano_huevo', category: 'Horneado', label: 'Baño de huevo', cost: 0.076 },
-  { id: 'gas_horno', category: 'Horneado', label: 'Gas horno', cost: 0.231 },
-
-  { id: 'papel_kraft', category: 'Envasado y Etiquetado', label: 'Papel kraft base', cost: 0.11 },
-  { id: 'bandeja_carton', category: 'Envasado y Etiquetado', label: 'Bandeja cartón', cost: 0.09 },
-  { id: 'sobre_papel', category: 'Envasado y Etiquetado', label: 'Sobre papel', cost: 0.028 },
-  { id: 'plastico_film', category: 'Envasado y Etiquetado', label: 'Plástico film', cost: 0.0025 },
-  { id: 'etiqueta_caja', category: 'Envasado y Etiquetado', label: 'Etiqueta caja', cost: 0.034 },
-  { id: 'folio_direccion', category: 'Envasado y Etiquetado', label: 'Folio etiqueta dirección', cost: 0.0043 },
-  { id: 'toner', category: 'Envasado y Etiquetado', label: 'Tóner impresora', cost: 0.0234 },
-  { id: 'cinta_adhesiva', category: 'Envasado y Etiquetado', label: 'Cinta adhesiva embalaje', cost: 0.001 },
-
-  { id: 'mano_de_obra', category: 'Mano de obra', label: 'Mano de obra', cost: 0.72 }
-]
-
-export default function Home() {
-  const [costs, setCosts] = useState<CostItem[]>(initialCosts.map(c => ({ ...c, isEditing: false })))
-  const [margin, setMargin] = useState<number>(0)
-  const [showTotals, setShowTotals] = useState<boolean>(false)
-  const [name, setName] = useState<string>('')
-  const [saved, setSaved] = useState<Empanada[]>([])
-  const [selected, setSelected] = useState<string>('')
-
-  useEffect(() => {
-    fetch('/api/empanadas')
-      .then(res => res.json())
-      .then(setSaved)
-      .catch(() => {})
-  }, [])
-
-  const handleCostChange = (id: string, value: number) => {
-    setCosts(costs.map(item => item.id === id ? { ...item, cost: value } : item))
-  }
-
-  const toggleEdit = (id: string) => {
-    setCosts(costs.map(item => item.id === id ? { ...item, isEditing: !item.isEditing } : item))
-  }
-
-  const saveEmpanada = async () => {
-    const payload: Empanada = { name, costs: costs.map(({ isEditing, ...rest }) => rest), margin }
-    await fetch('/api/empanadas', {
+  const handleSubmit = async (e: React.FormEvent) => {
+    e.preventDefault()
+    const res = await fetch('/api/login', {
       method: 'POST',
       headers: { 'Content-Type': 'application/json' },
-      body: JSON.stringify(payload)
+      body: JSON.stringify({ username, password })
     })
-    const list = await fetch('/api/empanadas').then(res => res.json())
-    setSaved(list)
+    if (res.ok) {
+      router.push('/calculadora')
+    } else {
+      alert('Credenciales inválidas')
+    }
   }
-
-  const loadEmpanada = (emp: Empanada) => {
-    setCosts(emp.costs.map(c => ({ ...c, isEditing: false })))
-    setMargin(emp.margin)
-    setShowTotals(false)
-  }
-
-  const total = costs.reduce((sum, item) => sum + (item.cost || 0), 0)
-  const vat = total * 0.10
-  const totalWithVat = total + vat
-  const sellingPrice = totalWithVat * (1 + (margin || 0) / 100)
-  const profit = sellingPrice - totalWithVat
-
-  const categories = Array.from(new Set(costs.map(c => c.category)))
 
   return (
-
-    <div className="p-4 font-sans max-w-screen-lg mx-auto bg-white rounded shadow">
-      <h1>Calculadora de Costes de Empanada de Carne</h1>
-
-      <div className="mb-4">
+    <div className="p-4 max-w-sm mx-auto">
+      <h1 className="text-xl font-bold mb-4">Login</h1>
+      <form onSubmit={handleSubmit} className="flex flex-col gap-2">
         <input
           type="text"
-          placeholder="Nombre de empanada"
-          value={name}
-          onChange={e => setName(e.target.value)}
+          placeholder="Usuario"
+          value={username}
+          onChange={e => setUsername(e.target.value)}
+          className="border p-2"
         />
-        <button onClick={saveEmpanada} className="ml-2">Guardar empanada</button>
-      </div>
-
-      <div className="mb-4">
-        <select value={selected} onChange={e => setSelected(e.target.value)}>
-          <option value="">Cargar empanada...</option>
-          {saved.map(emp => (
-            <option key={emp.name} value={emp.name}>{emp.name}</option>
-          ))}
-        </select>
-        <button
-          onClick={() => {
-            const emp = saved.find(e => e.name === selected)
-            if (emp) loadEmpanada(emp)
-          }}
-          className="ml-2"
-        >
-          Cargar
+        <input
+          type="password"
+          placeholder="Contraseña"
+          value={password}
+          onChange={e => setPassword(e.target.value)}
+          className="border p-2"
+        />
+        <button type="submit" className="bg-green-600 text-white py-2">
+          Entrar
         </button>
-      </div>
-
-
-      {categories.map(cat => (
-        <div key={cat} className="mb-4">
-          <h2 className="text-xl font-semibold mb-2">{cat}</h2>
-          <table className="w-full border-collapse table-auto">
-            <thead>
-              <tr className="text-left border-b">
-                <th className="pb-1">Concepto</th>
-                <th className="pb-1">Costo (€)</th>
-              </tr>
-            </thead>
-            <tbody>
-              {costs.filter(c => c.category === cat).map(item => (
-
-                <tr key={item.id}>
-                  <td>{item.label}</td>
-                  <td>
-                    {item.isEditing ? (
-                      <>
-                        <input
-                          type="number"
-                          value={item.cost}
-                          step="0.0001"
-                          onChange={e => handleCostChange(item.id, parseFloat(e.target.value))}
-                        />
-                        <button onClick={() => toggleEdit(item.id)} className="ml-2">Guardar</button>
-                      </>
-                    ) : (
-                      <>
-                        {item.cost.toFixed(4)}
-                        <button onClick={() => toggleEdit(item.id)} className="ml-2">Editar</button>
-                      </>
-                    )}
-
-                  </td>
-                </tr>
-              ))}
-            </tbody>
-          </table>
-        </div>
-      ))}
-
-      <div className="mt-4">
-        <label>
-          Margen de beneficio (%):
-
-          <input
-            className="border rounded px-2 py-1 w-24"
-            type="number"
-            value={margin}
-            step="0.01"
-            onChange={e => setMargin(parseFloat(e.target.value))}
-          />
-        </label>
-      </div>
-
-
-      <button onClick={() => setShowTotals(true)} className="mt-4">
-        Obtener gastos y beneficios
-      </button>
-
-      {showTotals && (
-        <div className="mt-4">
-          <p>Total: {total.toFixed(4)} €</p>
-          <p>IVA (10%): {vat.toFixed(4)} €</p>
-          <p>Total con IVA: {totalWithVat.toFixed(4)} €</p>
-          <p>Precio de venta: {sellingPrice.toFixed(4)} €</p>
-          <p>Beneficio: {profit.toFixed(4)} €</p>
-        </div>
-      )}
-
+      </form>
     </div>
   )
 }
