+++ package.json
@@ -10,7 +10,8 @@
   "dependencies": {
     "next": "14.0.0",
     "react": "18.2.0",
-    "react-dom": "18.2.0"
+    "react-dom": "18.2.0",
+    "mongodb": "^5.7.0"
   },
   "devDependencies": {
     "typescript": "^5.2.2",
